Turn the review request infobox into a class-based view.
This updates the new review request infobox and turns it into a
class-based view, taking advantage of the mixin for review request
access control. It also fixes up some of the code to work for anonymous
users and to exclude draft reviews and replies to reviews.

Testing Done:
Tested the review request infobox as an authenticated user and as a
normal user.

Tested with review requests containing draft reviews and replies.

Reviewed at https://reviews.reviewboard.org/r/9020/