Move the showSendEmail attribute into ReviewRequestEditor.
The `showSendEmail` attribute was being passed to `ChangeBoxView`,
`ReviewBoxView`, and `InitialStatusUpdatesBoxView` individually, and then
being forwarded on to other views. This state is really global, and not
per-box, and with some of the upcoming work, it ends up being
increasingly annoying to pass this state individually through the views.

The attribute is now part of `ReviewRequestEditor`, which gets passed to
these views instead. Some of the views already take this editor, and
more will need it with some upcoming work anyway. Some of the new
statements will even be rewritten in the upcoming change. This just
simplifies the options/attributes needed by views and lays the ground
work for those changes.

Testing Done:
JavaScript unit tests passed.

Tested the draft banner for review request updates, draft reviews on a
review request, draft replies on reviews, draft replies on initial status
updates, and draft replies on changedesc-associated status updates. These
were all tested with and without e-mail enabled (which is what the
`showSendEmail` flag is bound to).

Reviewed at https://reviews.reviewboard.org/r/9034/