Disable OAuth2 apps when their owner is removed from the Local Site
OAuth2 applications are now disabled when their owner is removed from
the Local Site the application is assigned to. The application will be
reassigned to one of the LocalSite administrators and set to `disabled`,
which will (once authentication is implemented) prevent the Application
from being used for authentication.

Testing Done:
- Ran unit tests.
- Ran the evolution.

Reviewed at https://reviews.reviewboard.org/r/9033/