Convert the diff comment fragments view to a class-based view.
This updates the diff comment fragments view to a class-based view,
making use of the mixins for handling review request access and ETags.

The logic is the same between the old view and new, with the exception
that we no longer set an Expires header. This hasn't really been needed
since we began using ETags for this view.

Testing Done:
Tested that diff comment fragments continued to load.

Unit tests passed.

Reviewed at https://reviews.reviewboard.org/r/9029/