Fix 500s from the diff context resource
The refactoring of views into mixins broke subtly broke the diff context
resource. It was passing in all arguments positionally and the new
CheckLocalSiteAccessViewMixin was pulling `local_site_name` out of both
kwargs and (mistakenly) from a caller passing `*args`. We now pass all
arguments in as keyword arguments to avoid this mess.

This patch also removes a bunch of dead imports I noticed in
`reviews/views` as a result of the aforementioned refactor.

Testing Done:
- Ran unit tests.
- With this patch applied, there are no more 500s on the diff context
  resource when viewing interdiffs.

Reviewed at https://reviews.reviewboard.org/r/9065/