Do not update timestamp when revoking Ship Its
Previously, revoking a Ship It would cause the timestamp of the review
to be updated, which resulted in the comments on a review request going
out of order. This was because we were not passing keyword arguments
from `Review.save` to its parent method. This is now fixed and unit
tests have been added to ensure the behaviour does not regress.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/9068/