Add new common CSS rules for infoboxes.
We had a mixture of styles for our infoboxes, causing the user, bug, and
review request infoboxes to have inconsistent padding and presentation.
This change introduces new rules (now living in `css/ui/infoboxes.less`)
that can be used for any kind of infobox. These define the infobox
itself, a list of links, a picture to display on the side, text
sections, and small detail information for the header.

The user infobox is the model for a lot of this, as its presentation is
the most finely-tuned at this point and was mostly using generic infobox
CSS class names. The bug and review request infoboxes have been updated
to use these class names as well.

The class names for the infoboxes themselves have been updated to use
hyphens instead of underscores, now that the names aren't bound to the
jQuery method names.

Presentation-wise, nothing has really changed except that the padding
around the review request and bug infoboxes now match that of the user
infobox. Upcoming changes will tweak the presentation for these infoboxes
further, better improving the look and consistency across infoboxes.

Testing Done:
Checked the display of all three infoboxes. Made sure the user infobox
did not regress in presentation, and that the other infoboxes looked a
bit better.

Reviewed at https://reviews.reviewboard.org/r/9067/