Take advantage of the new datagrid link-related features/fixes in Djblets.
Djblets just received some fixes for the `link=False` handling, along
with support for specifying the CSS classes for the generated links. We
now make use of these to reduce the data that must be rendered in the
user and summary columns, and to fix turning off links for the bugs
column.

Testing Done:
Tested that infoboxes and links still worked for review request summaries,
users, and bugs.

Bugs closed: 9080

Reviewed at https://reviews.reviewboard.org/r/9081/