Fix attaching infoboxes when editing Bugs, Reviewers, and Depends On fields.
Editing the Depends On field on a review request caused the generated
links to no longer have the CSS classes necessary to attach infoboxes.
The Bugs and Reviewers fields worked around this, but they had to
implement this themselves, and it wasn't consistent.

The function for creating a list of URLs for injecting into the field
(`urlizeList`) now takes an options parameter, instead of positional
arguments, for the building functions. One of those options is a CSS
class to apply to each link. The Bugs, Reviewers, and Depends On fields
now take advantage of this, receiving links that they can then easily
attach infoboxes to.

`urlizeList` also makes use of `convertToLink`, which is where the new
option support ultimately lives. This allows us to stay consistent
between the two link-building functions.

Both also now return jQuery elements, instead of building and returning
a string.

Testing Done:
Edited all affected fields, making sure the links built correctly with
various numbers of values. Checked that all CSS classes were applied and
that infoboxes worked.

Reviewed at https://reviews.reviewboard.org/r/9087/