Restyle the bug infobox.
Historically, the bug infobox has never looked all that amazing. It had
some content from the bug report, which was trimmed, and that sat below
the description and status, but none of it was presented well. Some of
the recent infobox worked helped with the rough edges (lack of padding
around the edges and bad default sizing), but it still wasn't enough to
render the content very nicely.

The new infobox works in combination with some enhancements to the base
bug tracker support. We now render the header of the infobox much like
the review request infobox, with the summary in the upper-left (and the
status below with an icon) and the bug ID in the upper-right. Below
that, we now render the bug text in the best way possible given the text
format provided.

Bug tracker implementations can now return the text format (`html`,
`markdown`, or `plain`). Based on the format, the infobox will attempt
to normalize the content for display, in a secure way. When it's
Markdown, we apply our Markdown processor and formatting rules. For
HTML, we strip tags, convert standard entities, and escape the rest.

The bug text is scrollable, like the new review request infobox's
description text. Hovering over the text with the mouse will allow the
user to read the contents of the bug description.

Down the road, we can extend this further by allowing additional
metadata from the bug tracker to be displayed.

The Splat implementation received updates for passing on the text format
for the infobox.

Testing Done:
Tested with bugs on GitHub and Splat.

On Splat, I tested with tickets marked as HTML, plain text, and Markdown,
and verified that each was normalized and formatted appropriately.

Reviewed at https://reviews.reviewboard.org/r/9088/