Fix OAuth API tests and add some helpful logging for the future.
The OAuth token resource tests were failing because the feature isn't yet
enabled by default. This change adds the necessary feature check override to
the tests (to ensure that it's enabled before calling it), and adds some
logging so that if we encounter this situation in the future, we'll see some
useful output instead of having to trace through the code to find out where the
`PERMISSION_DENIED` is coming from.

Testing done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/9109/