Prevent XSS code execution attacks through location.hash
We were not always consistent with treatment of the location hash. We
now treat it as always potentially hostile. If the original (unencoded)
hash contains any non-alphanumeric (or comma) characters, we will reject
it and treat it as if the hash were empty.

This fixes a security issue where the diff viewer and comment issue
table could inject executable JavaScript into the page through the
location.hash containing malicious content.

A unit test has been added to ensure the behaviour does not regress.

Testing Done:
- With the patch applied, I am no longer able to trigger the XSS
  vulnerability.
- Ran JS tests.
- Verified diff and interdiff navigation still works correctly.
- Verified diff pagination still works correctly.
- Verified issue table expands the comment box of an issue when
  an issue is clicked.

Reviewed at https://reviews.reviewboard.org/r/9104/