Change "box" terminology to "entry" for the review request page.
This updates the JavaScript for the review request page to use "entry"
instead of "box" for the terminology, helping to make the code a bit
more readable and clear. It also fixes up some of the event and anchor
handling to be more efficient.

Testing Done:
Tested that collapsing/expanding all and individual entries worked.

Tested entry expansion when anchors are referenced.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/9110/