Move and rename the entry-related templates.
In an effort to transition and clean up the terminology for entries on
the review request page, this change renames the `reviews/boxes/`
template directory to `reviews/entries/` and makes some changes for
what's inside that directory.

Includable utility templates used by some of the existing entry
templates are now prefixed with an underscore to designate that they are
not main templates themselves. This includes the review body for both
standard and status update reviews.

The base template for entries now lives in `reviews/entries/` as well,
renamed to `base.html`.

Testing Done:
Tested that all the entry templates and utility templates were all
working correctly. This includes reviews in the initial status updates
and review status updates.

Reviewed at https://reviews.reviewboard.org/r/9111/