Use importlib to check for elasticsearch instead of import statement
The `import elasticsearch` statement in `elasticsearch.py` was
succeeding if the module was not installed because it would find
`./elasticsearch.py` (itself) and import that instead. We now use
`import_module` to do the dependency check, which checks for
system-level modules and not local ones.

Testing Done:
With this patch applied, I cannot enable `elasticsearch` when it is not
installed.

Reviewed at https://reviews.reviewboard.org/r/9099/