Always include the Issue Summary Table on the page.
The Issue Summary Table was conditionally added to the review request
page only if there were one or more issues filed against the review
request. This is so we didn't unnecessarily show an empty table.
However, this also meant that we had to conditionally add the view for
this table (which we used to do but no longer do in 3.0, as of a recent
change). Not having the table in there also meant we had no ability to
dynamically update the table as new data came in for the page (which is
coming soon).

This change unconditionally adds the table to the page. It does,
however, hide it if there are no open issues, so it retains the same
behavior as it did before. Future changes will build upon this to show
the table when there are issues.

Testing Done:
Loaded a page without an Issue Summary Table and saw that no JavaScript
errors had appeared, and that the elements for the table was in the DOM
(but were hidden).

Loaded a page with issues and saw that the table was visible.

Reviewed at https://reviews.reviewboard.org/r/9123/