Smooth out some wrinkles in review request field rendering.
The way that review request fields were rendered was a bit hodge-podge. The
templatetags that iterated over fieldsets had a handful of special-cases that
theoretically simplified the template, but it wasn't really buying us anything.
In addition, we had three different implementations of rendering the actual
field HTML, including a completely hard-coded one for the summary.

This change generalizes everything a bit, adding conditionals in the template
for the special cases of the main fieldset and the summary field. This also
breaks out the actual field HTML into a reusable template. I've also changed
`should_render` to be a property on the field, which can be overridden as a
`@property` method in those cases that need it.

Testing done:
- Ran unit tests.
- Ran js-tests.
- Smoke tested review request fields.

Reviewed at https://reviews.reviewboard.org/r/9134/