Fix centering the collapse button in diff fragments.
When the code for handling the vertical centering of the collapse button
for diff fragments was moved into `RB.CenteredElementManager`, the
centering logic regressed. The new logic expected that the element would
center relative to a single container, but the old logic carefully
calculated a top and bottom container for the centering, allowing the
collapse button to appear correctly centered across several diff chunks.

`RB.CenteredElementManager` now allows for a top and bottom container to
be specified, and will use both of those for the centering logic. If
only a top container is specified, it will be used for both.

This fixes the display of the collapse button when expanding diffs.

Testing Done:
Expanded several diff fragments in both directions, repeatedly, and
saw that the collapse button was always properly vertically centered in
relation to the entire visible diff area, and to the screen.

Reviewed at https://reviews.reviewboard.org/r/9136/