Move comment anchor URL handling into the ReviewRequestPage.PageView.
Until now, the Issue Summary Table has been handling URLs with a
`#comment123` hash and taking care of expanding the appropriate review
and jumping to the right place on the page. This meant that these URLs
wouldn't work if the Issue Summary Table wasn't being used on the page,
and was a bit redundant with the page view's own URL handling.

This change moves the handling of anchors fully into the review request
page's view, consolidating logic into one place and making the Issue
Summary Table fully self-contained.

Testing Done:
Unit tests pass.

Tested URLs pointing to an issue in a collapsed entry, and saw that it
handled navigating to the right place.

Reviewed at https://reviews.reviewboard.org/r/9122/