Fix RB.DnDUploader state leakage between unit tests.
Many tests are at some point dependent on `RB.DnDUploader`. Some tests
initialize this, some clear it out, and some don't specifically deal
with it but views they set up do. Depending on which subsets of tests
are run, tests may break due to a lack of an available instance for the
uploader, or due to an instance not being cleaned up.

This change fixes up all those tests, ensuring that we always clear it
out in views that implicitly or explicitly deal with it, and create it
if necessary prior to running tests.

Testing Done:
All unit tests pass.

Ran subsets of tests that were breaking before when run in isolation,
and saw that they now pass.

Reviewed at https://reviews.reviewboard.org/r/9145/